requirements.txt
streamlit
openai>=1.2.0
langchain>=0.1.14
langchain-community>=0.0.21
langchain-openai>=0.0.8
PyPDF2
faiss-cpu
tiktoken
python-dotenv
numexpr
